# Conflicts:
#	identity/src/main/kotlin/net/thechance/identity/api/controller/IdentityController.kt
#	identity/src/main/kotlin/net/thechance/identity/service/ResetPasswordService.kt
#	identity/src/main/kotlin/net/thechance/identity/service/UserService.kt
#	identity/src/test/kotlin/net/thechance/identity/service/ResetPasswordServiceTest.kt

+++ trends/src/main/kotlin/net/thechance/trends/service/ReelsService.kt
@@ -1,8 +1,9 @@
 package net.thechance.trends.service
 
-import net.thechance.trends.api.controller.exception.ReelNotFoundException
-import net.thechance.trends.api.controller.exception.TrendCategoryNotFoundException
 import net.thechance.trends.entity.Reel
+import net.thechance.trends.exception.ReelNotFoundException
+import net.thechance.trends.exception.TrendCategoryNotFoundException
+import net.thechance.trends.exception.VideoDeleteFailedException
 import net.thechance.trends.repository.CategoryRepository
 import net.thechance.trends.repository.ReelsRepository
 import org.springframework.data.domain.Page
@@ -11,12 +12,14 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import org.springframework.web.multipart.MultipartFile
 import java.util.*
 
 @Service
 class ReelsService(
     private val reelsRepository: ReelsRepository,
-    private val categoryRepository: CategoryRepository
+    private val categoryRepository: CategoryRepository,
+    private val videoStorageService: VideoStorageService,
 ) {
     fun getAllReelsByUserId(
         pageable: Pageable,
@@ -34,12 +37,16 @@ class ReelsService(
         return body
     }
 
+    @Transactional
     fun deleteReelById(id: UUID, currentUserId: UUID) {
-        val isReelExists = reelsRepository.existsByIdAndOwnerId(id , currentUserId)
+        val reel = reelsRepository.findByIdAndOwnerId(id, currentUserId)
+            ?: throw ReelNotFoundException()
 
-        if (!isReelExists) throw ReelNotFoundException()
-
-        reelsRepository.deleteById(id)
+        runCatching {
+            if (videoStorageService.deleteVideo(reel.videoUrl)) reelsRepository.deleteById(id)
+        }.onFailure {
+            throw VideoDeleteFailedException()
+        }
     }
 
     @Transactional
@@ -61,5 +68,23 @@ class ReelsService(
         )
 
         return reelsRepository.save(updatedReel)
+    }
+
+
+    fun uploadReel(currentUserId: UUID, file: MultipartFile): UUID {
+        val videoUrl = videoStorageService.uploadVideo(
+            file = file,
+            fileName = file.originalFilename ?: "Untitled",
+            folderName = TRENDS_FOLDER_NAME
+        )
+        val reel = Reel(
+            ownerId = currentUserId,
+            videoUrl = videoUrl
+        )
+        return reelsRepository.save(reel).id
+    }
+
+    companion object {
+        private const val TRENDS_FOLDER_NAME = "trends"
     }
 }
